Recator:Separated each function into its own file to improve code organization and readability
